+++ tests/unit/plugins/modules/test_nso_action.py
@@ -22,7 +22,7 @@ __metaclass__ = type
 import json
 
 from ansible_collections.cisco.nso.tests.unit.compat.mock import patch
-from ansible_collections.cisco.nso.plugins.modules.network.nso import nso_action
+from ansible_collections.cisco.nso.plugins.modules import nso_action
 from . import nso_module
 from .nso_module import MockResponse